Move wiki tests into test directory
